Result of integer multiplication cast to long in org.molgenis.framework.server.services.CleanTmpDirProcess.run()

+++ src/org/molgenis/framework/server/services/MolgenisCleanTmpDirService.java
@@ -42,10 +42,10 @@ public class MolgenisCleanTmpDirService implements MolgenisService
 
 class CleanTmpDirProcess implements Runnable
 {
-	private int mfa;
-	private int hox;
+	private long mfa;
+	private long hox;
 
-	CleanTmpDirProcess(int howOftenExecutedInSeconds, int maxFileAgeInHours)
+	CleanTmpDirProcess(int howOftenExecutedInSeconds, long maxFileAgeInHours)
 	{
 		hox = howOftenExecutedInSeconds;
 		mfa = maxFileAgeInHours;
